# docs/swagger.yaml
openapi: 3.0.2
info:
    title: Games-API-Rogovski
    version: 1.0.0
    description: '### This is a Swagger UI for our games API'
paths:
    /games:
        get:
            tags:
                - games
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                type: string
                            examples:
                                List of games:
                                    value:
                                        - Witcher 3
                                        - Minecraft
                    description: Returns a list of games
            summary: Get a list of all games
            description: |
                Returns a list of all games in API
        post:
            requestBody:
                description: A new game to be created
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/error'
                        examples:
                            Game /w+ was not added:
                                value:
                                    error: One or all params are missing
                required: true
            tags:
                - new-game
            responses:
                '201':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/new-game'
                    description: Succesfully created
                '400':
                    description: 'One or all params are missing. '
            summary: Creates a game
            description: Creates a new game
    '/games/{id}':
        get:
            tags:
                - details
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                type: string
                            examples:
                                Full details of a given game:
                                    value:
                                        id: 1
                                        name: Witcher 3
                                        price: 29.99
                    description: |
                        Returns full details of a fiven id game
                '404':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/error'
                            examples:
                                Game /w+ was not found:
                                    value:
                                        error: Game \w+ was not found in game array
                    description: Game was not found in games array
            operationId: id
            summary: Get full details of a game
            description: Returns full details of a given id game
        delete:
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/delete-game'
            tags:
                - delete-game
            responses:
                '204':
                    description: No content
                '404':
                    description: |
                        No game with this id
            summary: Delete a game
            description: Delete a choosed game
components:
    schemas:
        error:
            title: Root Type for error
            description: ''
            type: object
            properties:
                error:
                    type: string
            example:
                error: Error message.
        new-game:
            title: Root Type for new-game
            description: ''
            required:
                - price
            type: object
            properties:
                name:
                    type: string
                price:
                    format: float
                    type: number
            example:
                name: Tetris
                price: 0.99
        delete-game:
            title: Root Type for delete-game
            description: ''
            required:
                - id
            type: object
            properties:
                id:
                    type: string
            example:
                id: 1